Migrate plugin to the structure used since Shopware v5.2
