+++ Resources/services.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" ?>
+
+<container xmlns="http://symfony.com/schema/dic/services"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
+
+    <services>
+        <service id="swag_media_sftp.subscriber.adapter_collection_subscriber" class="SwagMediaSftp\Subscriber\AdapterCollectionSubscriber">
+            <tag name="shopware.event_subscriber" />
+        </service>
+    </services>
+
+</container>
