+++ SwagMediaSftp.php
@@ -1,0 +1,46 @@
+<?php
+/*
+ * (c) shopware AG <info@shopware.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ */
+
+namespace SwagMediaSftp;
+
+require __DIR__ . "/vendor/autoload.php";
+
+use Shopware\Components\Plugin;
+use Shopware\Components\Plugin\Context\ActivateContext;
+use Shopware\Components\Plugin\Context\DeactivateContext;
+use Shopware\Components\Plugin\Context\InstallContext;
+use Shopware\Components\Plugin\Context\UninstallContext;
+use Shopware\Components\Plugin\Context\UpdateContext;
+
+class SwagMediaSftp extends Plugin
+{
+    public function update(UpdateContext $context)
+    {
+        $context->scheduleClearCache(InstallContext::CACHE_LIST_DEFAULT);
+        parent::update($context);
+    }
+
+    public function activate(ActivateContext $context)
+    {
+        $context->scheduleClearCache(InstallContext::CACHE_LIST_DEFAULT);
+        parent::activate($context);
+    }
+
+    public function deactivate(DeactivateContext $context)
+    {
+        $context->scheduleClearCache(InstallContext::CACHE_LIST_DEFAULT);
+        parent::deactivate($context);
+    }
+
+    public function uninstall(UninstallContext $context)
+    {
+        $context->scheduleClearCache(InstallContext::CACHE_LIST_DEFAULT);
+        parent::uninstall($context);
+    }
+}
