+++ plugin.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<plugin xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="https://raw.githubusercontent.com/shopware/shopware/5.2/engine/Shopware/Components/Plugin/schema/plugin.xsd">
+
+    <label lang="de">Media Adapter: SFTP</label>
+    <label lang="en">Media Adapter: SFTP</label>
+
+    <version>1.0.0</version>
+    <author>shopware AG</author>
+    <license>MIT</license>
+    <description>SFTP-Erweiterung für die Media Adapter</description>
+
+    <compatibility minVersion="5.2.0" />
+
+</plugin>
